+++ packages/stream_core_flutter/lib/src/components/reaction/stream_reaction_picker.dart
@@ -210,8 +210,8 @@ class DefaultStreamReactionPicker extends StatelessWidget {
             size: .small,
             type: .outline,
             style: .secondary,
-            icon: icons.plus,
-            onTap: props.onAddReactionTap,
+            icon: Icon(icons.plus),
+            onPressed: props.onAddReactionTap,
           ),
         ],
       ),
@@ -234,7 +234,7 @@ class _StreamReactionPickerThemeDefaults extends StreamReactionPickerThemeData {
   double get elevation => 3;
 
   @override
-  double get spacing => _spacing.none;
+  double get spacing => _spacing.xxxs;
 
   @override
   Color get backgroundColor => _colorScheme.backgroundElevation2;